Clean up code and readme.
Messages are now in META-INF

+++ src/main/java/net/trajano/wagon/git/GitWagon.java
@@ -1,5 +1,7 @@
 package net.trajano.wagon.git;
 
+import static java.lang.String.format;
+
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
@@ -7,6 +9,7 @@ import java.io.IOException;
 import java.net.URISyntaxException;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.ResourceBundle;
 
 import org.apache.maven.wagon.ConnectionException;
 import org.apache.maven.wagon.InputData;
@@ -30,6 +33,12 @@ import org.eclipse.jgit.transport.UsernamePasswordCredentialsProvider;
  */
 @Component(role = Wagon.class, hint = "git", instantiationStrategy = "per-lookup")
 public class GitWagon extends StreamWagon {
+
+    /**
+     * Resource bundle.
+     */
+    private static final ResourceBundle R = ResourceBundle
+            .getBundle("META-INF/Messages"); //$NON-NLS-1$
 
     /**
      * Credentials provider.
@@ -59,8 +68,8 @@ public class GitWagon extends StreamWagon {
     @Override
     public void closeConnection() throws ConnectionException {
         try {
-            git.add().addFilepattern(".").call();
-            git.commit().setMessage("wagon-git commit").call();
+            git.add().addFilepattern(".").call(); //$NON-NLS-1$
+            git.commit().setMessage(R.getString("commitmessage")).call(); //$NON-NLS-1$
             git.push().setRemote(gitUri.getGitRepositoryUri())
                     .setCredentialsProvider(credentialsProvider).call();
         } catch (final GitAPIException e) {
@@ -70,7 +79,12 @@ public class GitWagon extends StreamWagon {
 
     /**
      * This will read from the working copy. File modification date would not be
-     * available as it does not really have any meaningful value.{@inheritDoc}
+     * available as it does not really have any meaningful value. {@inheritDoc}
+     * 
+     * @throws ResourceDoesNotExistException
+     *             when the file does not exist
+     * @throws AuthorizationException
+     *             when the file cannot be read
      */
     @Override
     public void fillInputData(final InputData inputData)
@@ -80,8 +94,12 @@ public class GitWagon extends StreamWagon {
             final File file = new File(gitDir, inputData.getResource()
                     .getName());
             if (!file.exists()) {
-                throw new ResourceDoesNotExistException("The resource " + file
-                        + " does not exist.");
+                throw new ResourceDoesNotExistException(format(
+                        R.getString("filenotfound"), file)); //$NON-NLS-1$
+            }
+            if (!file.canRead()) {
+                throw new AuthorizationException(format(
+                        R.getString("cannotreadfile"), file)); //$NON-NLS-1$
             }
             inputData.setInputStream(new FileInputStream(file));
             inputData.getResource().setContentLength(file.length());
@@ -99,13 +117,21 @@ public class GitWagon extends StreamWagon {
         try {
             final File file = new File(gitDir, outputData.getResource()
                     .getName());
-            file.getParentFile().mkdirs();
+            if (!file.getParentFile().mkdirs()
+                    && !file.getParentFile().exists()) {
+                throw new TransferFailedException(format(
+                        R.getString("unabletocreatedirs"), //$NON-NLS-1$
+                        file.getParentFile()));
+            }
             outputData.setOutputStream(new FileOutputStream(file));
         } catch (final IOException e) {
             throw new TransferFailedException(e.getMessage(), e);
         }
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public List<String> getFileList(final String destinationDirectory)
             throws TransferFailedException, ResourceDoesNotExistException,
@@ -113,14 +139,14 @@ public class GitWagon extends StreamWagon {
         final File dir = new File(gitDir, destinationDirectory);
         final File[] files = dir.listFiles();
         if (files == null) {
-            throw new ResourceDoesNotExistException("The directory "
-                    + destinationDirectory + " does not exist");
+            throw new ResourceDoesNotExistException(format(
+                    R.getString("dirnotfound"), dir)); //$NON-NLS-1$
         }
         final List<String> list = new LinkedList<String>();
         for (final File file : files) {
             String name = file.getName();
-            if (file.isDirectory() && !name.endsWith("/")) {
-                name += "/"; // NOPMD this is easier to read.
+            if (file.isDirectory() && !name.endsWith("/")) { //$NON-NLS-1$
+                name += "/"; // NOPMD this is easier to read. //$NON-NLS-1$
             }
             list.add(name);
         }
@@ -136,13 +162,13 @@ public class GitWagon extends StreamWagon {
             AuthenticationException {
         try {
             gitUri = new GitUri(getRepository().getUrl());
-            gitDir = File.createTempFile("wagon-git", null);
+            gitDir = File.createTempFile("wagon-git", null); //$NON-NLS-1$
             gitDir.delete();
             gitDir.mkdir();
 
             credentialsProvider = new UsernamePasswordCredentialsProvider(
                     getAuthenticationInfo().getUserName(),
-                    getAuthenticationInfo().getPassword() == null ? ""
+                    getAuthenticationInfo().getPassword() == null ? "" //$NON-NLS-1$
                             : getAuthenticationInfo().getPassword());
             git = Git.cloneRepository().setURI(gitUri.getGitRepositoryUri())
                     .setCredentialsProvider(credentialsProvider)
@@ -152,7 +178,7 @@ public class GitWagon extends StreamWagon {
                 final RefUpdate refUpdate = git.getRepository()
                         .getRefDatabase().newUpdate(Constants.HEAD, true);
                 refUpdate.setForceUpdate(true);
-                refUpdate.link("refs/heads/" + gitUri.getBranchName());
+                refUpdate.link("refs/heads/" + gitUri.getBranchName()); //$NON-NLS-1$
             }
         } catch (final GitAPIException e) {
             throw new ConnectionException(e.getMessage(), e);
@@ -163,11 +189,18 @@ public class GitWagon extends StreamWagon {
         }
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public void putDirectory(final File sourceDirectory,
             final String destinationDirectory) throws TransferFailedException,
             ResourceDoesNotExistException, AuthorizationException {
         try {
+            if (!sourceDirectory.isDirectory()) {
+                throw new ResourceDoesNotExistException(format(
+                        R.getString("dirnotfound"), sourceDirectory)); //$NON-NLS-1$
+            }
             FileUtils.copyDirectoryStructure(sourceDirectory, new File(gitDir,
                     destinationDirectory));
         } catch (final IOException e) {
@@ -175,18 +208,25 @@ public class GitWagon extends StreamWagon {
         }
     }
 
+    /**
+     * {@inheritDoc} Does not throw anything.
+     */
     @Override
-    public boolean resourceExists(final String resourceName)
-            throws TransferFailedException, AuthorizationException {
+    public boolean resourceExists(final String resourceName) {
         final File file = new File(gitDir, resourceName);
 
-        if (resourceName.endsWith("/")) {
+        if (resourceName.endsWith("/")) { //$NON-NLS-1$
             return file.isDirectory();
         }
 
         return file.exists();
     }
 
+    /**
+     * Directory copy is supported.
+     * 
+     * @return true
+     */
     @Override
     public boolean supportsDirectoryCopy() {
         return true;
